requirements.txt
flask
requests
python-dotenv
langchain
langchain-groq
google-cloud-storage